Support iterables per ECMAScript 2015 spec.
In addition to "array like" elements, the ECMAScript specification
published June 2015 also supports "iterable" elements that have a
Symbol.iterator property. This commit adds support iterable
elements, closely matching the steps outlined at
http://www.ecma-international.org/ecma-262/6.0/#sec-array.from

A notable omission, however, is support for calling "[IteratorClose]"
in case of "[abrupt completion]" if an error is thrown by the mapping
function. It'd be lovely to implement support for that as well at
some point.

[IteratorClose]: http://www.ecma-international.org/ecma-262/6.0/#sec-iteratorclose
[Abrupt Completion]: http://www.ecma-international.org/ecma-262/6.0/#sec-completion-record-specification-type